% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{add_covariates}
\alias{add_covariates}
\title{Adds covariate data to sf object}
\usage{
add_covariates(
  grid_data,
  cov_data,
  zcols,
  weight_type = "area",
  popdens = NULL,
  verbose = TRUE,
  t_label = NULL
)
}
\arguments{
\item{grid_data}{sf object describing a regular grid over the area of interest.
See \link[rts2]{create_grid}}

\item{cov_data}{sf object or data.frame. See details.}

\item{zcols}{vector of character strings with the names of the columns of \code{cov_data}
to include}

\item{weight_type}{character string. Either "area" for area-weighted average or "pop"
for population-weighted average}

\item{popdens}{character string. The name of the column in \code{cov_data} with the
population density. Required if weight_type="pop"}

\item{verbose}{logical. Whether to provide a progress bar}

\item{t_label}{integer. If adding spatio-temporally varying data by time period,
this time label should be appended to the column name. See details.}
}
\value{
sf object identical to grid_data but with additional columns added with covariate data. See details.
}
\description{
Maps spatial, temporal, or spatio-temporal covariate data onto sf object
}
\details{
\emph{Spatially-varying data only} \verb{cov_data`` is an sf object describing covariate values for a set of polygons over the area of interest. The values are mapped onto another sf object - the grid }grid_data\verb{covering the same area (see \link[rts2]\{create_grid\}). For each grid cell in}grid_data\verb{a weighted average of each covariate listed in}zcols\verb{is generated with weights either equal to the area of intersection of the grid cell and the polygons in}cov_data\verb{ (}weight_type="area"\verb{), or this area multiplied by the population density of the polygon for population weighted (}weight_type="pop"\verb{). Columns with the names in }zcols` are added to the output.

\emph{Temporally-varying only data} \code{cov_data} is a data frame with number of rows
equal to the number of time periods. One of the columns must be called \code{t} and
have values from 1 to the number of time periods. The other columns of the data
frame have the values of the covariates for each time period. See
\link[rts2]{get_dow} for day of week data. A total of
length(zcols)*(number of time periods) columns are added to the output: for each
covariate there will be columns appended with each time period number. For example,
\code{dayMon1}, \code{dayMon2}, etc.

\emph{Spatially and temporally varying data} There are two ways to add data that
vary both spatially and temporally. The final output for use in analysis must
have a column for each covariate and each time period with the same name appended
by the time period number, e.g. \code{covariateA1},\code{covariateA2},... If the covariate
values for different time periods are in separate sf objects, one can follow
the method for spatially-varying only data above and append the time period number
using the argument \code{t_label}. If the values for different time periods are in the same
sf object then they should be named as described above and then can be added
as for spatially-varying covariates, e.g. \code{zcols=c("covariateA1","covariateA2")}.
}
\examples{
b1 <-  sf::st_sf(sf::st_sfc(sf::st_polygon(list(cbind(c(0,3,3,0,0),c(0,0,3,3,0))))))
g1 <- create_grid(b1,0.5)
cov1 <- create_grid(b1,0.8)
cov1$cov <- runif(nrow(cov1))
g1 <- add_covariates(g1,
                     cov1,
                     zcols="cov",
                    verbose = FALSE)
}
