% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_functions.R
\name{aggregate_output}
\alias{aggregate_output}
\title{Aggregate output}
\usage{
aggregate_output(
  grid_data,
  new_geom,
  zcols,
  weight_type = "area",
  popdens = NULL,
  verbose = TRUE
)
}
\arguments{
\item{grid_data}{sf object describing a regular grid over the area of interest.
Should be the same object used in the model fit. See \link[rts2]{create_grid} and
\link[rts2]{lgcp_fit}}

\item{new_geom}{sf object. A set of polygons covering the same area as \code{grid_data}}

\item{zcols}{vector of character strings. Names of the variables in \code{grid_data} to
map to the new geography}

\item{weight_type}{character string, either "area" or "pop" for area-weighted
or population weighted averaging, respectively}

\item{popdens}{character string. If \code{weight_type} is equal to "pop" then the
name of the column in \code{grid_data} with population density data}

\item{verbose}{logical. Whether to provide progress bar.}
}
\value{
An \code{sf} object identical to \code{new_geom} with additional columns with the
variables specified in \code{zcols}
}
\description{
Aggregate \code{lgcp_fit} output to another geography
}
\examples{
\dontrun{
b1 <- sf::st_sf(sf::st_sfc(sf::st_polygon(list(cbind(c(0,3,3,0,0),c(0,0,3,3,0))))))
g1 <- create_grid(b1,0.5)
dp <- data.frame(y=runif(10,0,3),x=runif(10,0,3),date=paste0("2021-01-",11:20))
dp <- create_points(dp,pos_vars = c('y','x'),t_var='date')
cov1 <- create_grid(b1,0.8)
cov1$cov <- runif(nrow(cov1))
g1 <- add_covariates(g1,
                     cov1,
                     zcols="cov",
                    verbose = FALSE)
g1 <- points_to_grid(g1, dp, laglength=5)
priors <- list(
  prior_lscale=c(0,0.5),
  prior_var=c(0,0.5),
  prior_linpred_mean=c(0),
  prior_linpred_sd=c(5)
  )
res <- lgcp_fit(g1,
                popdens="cov")
o1 <- extract_preds(g1,
                    res,
                    type=c("pred","rr"),
                    popdens="cov")
new1 <- aggregate_output(o1
                         cov1,
                         zcols="rr")
}
}
