% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\docType{data}
\name{birmingham_crime}
\alias{birmingham_crime}
\title{Birmingham crime data}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 132 rows and 21 columns.
}
\usage{
birmingham_crime
}
\description{
Counts of burglaries for the months of 2022 for the city of Birmingham, UK at the
Middle-Layer Super Output Area.
}
\keyword{datasets}
