% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\docType{data}
\name{boundary}
\alias{boundary}
\title{Boundary polygon for Birmingham, UK}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 1 rows and 2 columns.
}
\usage{
boundary
}
\description{
A Boundary polygon describing the border of the city of Birmingham, UK.
}
\keyword{datasets}
