% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_mcmlrts.R
\name{coef.rtsFit}
\alias{coef.rtsFit}
\title{Extracts fixed effect coefficients from a rtsFit object}
\usage{
\method{coef}{rtsFit}(object, ...)
}
\arguments{
\item{object}{An \code{rtsFit} model fit.}

\item{...}{Further arguments passed from other methods}
}
\value{
A named vector.
}
\description{
Extracts the fitted fixed effect coefficients from an \code{rtsFit} object returned from a call of \code{rtsFit} or \code{LA} in the \link[glmmrBase]{Model} class.
}
