% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_mcmlrts.R
\name{confint.rtsFit}
\alias{confint.rtsFit}
\title{Fixed effect confidence intervals for a \code{rtsFit} object}
\usage{
\method{confint}{rtsFit}(object, ...)
}
\arguments{
\item{object}{A \code{rtsFit} object.}

\item{...}{Further arguments passed from other methods}
}
\value{
A matrix (or vector) with columns giving lower and upper confidence limits for each parameter.
}
\description{
Returns the computed confidence intervals for a \code{rtsFit} object.
}
