% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_mcmlrts.R
\name{covariance.parameters}
\alias{covariance.parameters}
\title{Extracts the estimates of the covariance parameters}
\usage{
covariance.parameters(object)
}
\arguments{
\item{object}{An \code{mcml} model fit.}
}
\value{
A matrix of dimension (number of fixed effects ) x (number of MCMC samples). For Laplace approximation, the number of "samples" equals one.
}
\description{
Extracts the estimates of the covariance parameters an \verb{rtsFit object} returned from call of \code{lgcp_ml()} or \code{lgcp_bayes()} in the \link[rts2]{grid} class.
}
