% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{create_grid}
\alias{create_grid}
\title{Create sf grid over a boundary}
\usage{
create_grid(boundary, cellsize)
}
\arguments{
\item{boundary}{An sf object containing one polygon describing the area of interest}

\item{cellsize}{The dimension of the grid cells}
}
\value{
A sf object describing a set of square cells tiled over the area of the boundary
}
\description{
Produces a regular grid over an area of interest
}
\details{
Given a contiguous boundary describing an area of interest, returns an sf
object of a regular grid within the limits of the boundary.
}
\examples{
b1 = st_sf(st_sfc(st_polygon(list(cbind(c(0,3,3,0,0),c(0,0,3,3,0))))))
create_grid(b1,0.5)
}
