% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disaggregate.R
\name{disaggregate_covariate}
\alias{disaggregate_covariate}
\title{Disaggregate regional means to fine grid, preserving aggregate means}
\usage{
disaggregate_covariate(
  x_r,
  W,
  adj,
  pop = NULL,
  lambda_smooth = 1,
  lambda_ridge = 1e-04
)
}
\arguments{
\item{x_r}{length m vector of regional means (or pop-weighted means)}

\item{W}{m x n aggregation matrix (1 if cell in region, 0 otherwise)}

\item{adj}{n x n adjacency matrix}

\item{pop}{optional length n population vector; if provided, preserves
population-weighted means rather than simple means}

\item{lambda_smooth}{weight on spatial smoothness}

\item{lambda_ridge}{small ridge for numerical stability (needed because L is singular)}
}
\value{
A list
}
\description{
Disaggregate regional means to fine grid, preserving aggregate means
}
