% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\docType{data}
\name{example_points}
\alias{example_points}
\title{Simulated point data for running single-period examples}
\format{
An object of class \code{data.frame} with 261 rows and 3 columns.
}
\usage{
example_points
}
\description{
A set of 261 points simulated within the boundary of the city Birmingham, UK
from a log-Gaussian Cox process.
}
\keyword{datasets}
