% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_functions.R
\name{extract_preds}
\alias{extract_preds}
\title{Extract predictions}
\usage{
extract_preds(
  grid_data,
  stan_fit,
  type = c("pred", "rr", "irr"),
  irr.lag = NULL,
  t.lag = 0,
  popdens = NULL
)
}
\arguments{
\item{grid_data}{sf object describing a regular grid over the area of interest.
Should be the same object used in the model fit. See \link[rts2]{create_grid}}

\item{stan_fit}{A \link[rstan]{stanfit} or \link[cmdstanr]{CmdStanMCMC} object.
Output of \link[rts2]{lgcp_fit}}

\item{type}{Vector of character strings. Any combination of "pred", "rr", and "irr", which are,
posterior mean incidence (overall and population standardised), relative risk,
and incidence rate ratio, respectively.}

\item{irr.lag}{integer. If "irr" is requested as \code{type} then the number of time
periods lag previous the ratio is in comparison to}

\item{t.lag}{integer. Extract predictions for previous time periods.}

\item{popdens}{character string. Name of the column in \code{grid_data} with the
population density data}
}
\value{
An \code{sf} object identical to \code{grid_data} but with extra columns. See Details
}
\description{
Extract incidence and relative risk predictions
}
\details{
Three outputs can be extracted from the model fit, which will be reported
in the columns:

Predicted incidence: If type includes \code{pred} then \code{pred_mean_total} and
\code{pred_mean_total_sd} provide the
predicted mean total incidence and its standard deviation, respectively.
\code{pred_mean_pp} and \code{pred_mean_pp_sd} provide the predicted population
standardised incidence and its standard deviation.

Relative risk: if type includes \code{rr} then the relative risk is reported in
the columns \code{rr} and \code{rr_sd}. The relative risk here is the exponential
of the latent field, which describes the relative difference between
expexted mean and predicted mean incidence.

Incidence risk ratio: if type includes \code{irr} then the incidence rate ratio (IRR)
is reported in the columns \code{irr} and \code{irr_sd}. This is the ratio of the predicted
incidence in the last period (minus \code{t_lag}) to the predicted incidence in the
last period minus \code{irr_lag} (minus \code{t_lag}). For example, if the time period
is in days then setting \code{irr_lag} to 7 and leaving \code{t_lag=0} then the IRR
is the relative change in incidence in the present period compared to a week
prior.
}
