% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_functions.R
\name{extract_priors}
\alias{extract_priors}
\title{Generate priors from previous model fit}
\usage{
extract_priors(stan_fit)
}
\arguments{
\item{stan_fit}{A \link[rstan]{stanfit} or \link[cmdstanr]{CmdStanMCMC} object.
Output of \link[rts2]{lgcp_fit}}
}
\value{
list
}
\description{
Generate prior distributions from posterior distributions of previous model
fit by \code{lgcp_fit}
}
