% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_mcmlrts.R
\name{family.rtsFit}
\alias{family.rtsFit}
\title{Extracts the family from a \code{rtsFit} object.}
\usage{
\method{family}{rtsFit}(object, ...)
}
\arguments{
\item{object}{A \code{rtsFit} object.}

\item{...}{Further arguments passed from other methods}
}
\value{
A \link[stats]{family} object.
}
\description{
Extracts the \link[stats]{family} from a \code{rtsFit} object.
}
