% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_mcmlrts.R
\name{fitted.rtsFit}
\alias{fitted.rtsFit}
\title{Fitted values from a \code{rtsFit} object}
\usage{
\method{fitted}{rtsFit}(object, ...)
}
\arguments{
\item{object}{A \code{rtsFit} object.}

\item{...}{Further arguments passed from other methods}
}
\value{
Nothing, called for effects.
}
\description{
Fitted values should not be generated directly from an \code{rtsFit} object, rather fitted values should be
generated using the original \code{grid} object. A message is printed to the user.
}
