% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_mcmlrts.R
\name{fixed.effects}
\alias{fixed.effects}
\title{Extracts the fixed effect estimates}
\usage{
fixed.effects(object)
}
\arguments{
\item{object}{An \code{mcml} model fit.}
}
\value{
A named, numeric vector of fixed-effects estimates.
}
\description{
Extracts the fixed effect estimates from an \verb{rtsFit object} returned from call of \code{lgcp_ml()} or \code{lgcp_bayes()} in the \link[rts2]{grid} class.
}
