% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disaggregate.R
\name{flat_disaggregate}
\alias{flat_disaggregate}
\title{Simple flat disaggregation}
\usage{
flat_disaggregate(x_r, W)
}
\arguments{
\item{x_r}{Covariate to be disaggregtated}

\item{W}{weight matrix for disaggregation}
}
\value{
Vector
}
\description{
Simply solves t(W) \%*\% x_r
}
