% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_mcmlrts.R
\name{formula.grid}
\alias{formula.grid}
\title{Extracts the formula from a \code{grid} object.}
\usage{
\method{formula}{grid}(x, ...)
}
\arguments{
\item{x}{A \code{grid} object.}

\item{...}{Further arguments passed from other methods}
}
\value{
A \link[stats]{formula} object.
}
\description{
Extracts the \link[stats]{formula} from a \code{rtsFit} object stored in a grid object. Only returns the top level formula. For region models
this is the formula at the region level, otherwise the grid-level formula is returned. No random effects
specifications are included in the returned formula.
}
