% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{full_A}
\alias{full_A}
\title{Coverts a collapsed matrix A to full matrix A}
\usage{
full_A(A, NN)
}
\arguments{
\item{A}{A "collapsed" matrix A output from \link[rts2]{get_AD}}

\item{NN}{A matrix of integers with rows equal to the number of nearest neighbours and columns equal to the number of observations.
Each element is the set of nearest neighbours. See \link[rts2]{genNN}.}
}
\value{
A matrix
}
\description{
The output of the function \link{rts2}\link{get_AD} is a m x n matrix A, where
m is the number of nearest neighbours. This function coverts A into a full
n x n matrix.
}
