% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{genNN}
\alias{genNN}
\title{Generate matrix of nearest neighbours}
\usage{
genNN(x, M)
}
\arguments{
\item{x}{Matrix with x and y coordinates of the observations}

\item{M}{Number of nearest neighbours}
}
\value{
A M by n matrix of nearest neighbours. Note the indexing of the observations returned
by this function starts at zero.
}
\description{
Generates a M by n matrix of nearest neighbours. For observations with index less than M,
only i-1 nearest neighbours are generated.
}
