% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_AD}
\alias{get_AD}
\title{Generate the matrices A and D for the nearest neighbour approximation}
\usage{
get_AD(cov, data, NN, theta)
}
\arguments{
\item{cov}{An integer matrix with columns of block identifier, dimension of block, function definition, number of variables
in the argument to the funciton, and index of the parameters, respectively. Rows are specific functions of each block.}

\item{data}{Vector of data. Created by flattening the matrices in column-major order of the data used in each block.}

\item{NN}{A matrix of integers with rows equal to the number of nearest neighbours and columns equal to the number of observations.
Each element is the set of nearest neighbours. See \link[rts2]{genNN}.}

\item{gamma}{Vector of parameters used to generate the matrix D.}
}
\value{
A named list containing matrices A and D.
}
\description{
Generate the matrices A and D for the nearest neighbour approximation
}
\details{
The neatest neighbour Gaussian process approximation to the inverse covariance matrix
is \eqn{(I-A)D^-1(I-A)^T}. This function generates matrix D and a "collapsed" version of matrix A.
If there are \code{m} nearest neighbours then A will have \code{m} rows and columns equal to the number of observations, \code{n}.
The "full" A matrix is n x n. See \link[rts2]{full_A} for converting A to full representation.
}
