% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{get_dow}
\alias{get_dow}
\title{Generate day of week data}
\usage{
get_dow(grid_data)
}
\arguments{
\item{grid_data}{sf object with columns \code{t[0-9]} and \code{date[0-9]}. See \link[rts2]{points_to_grid}}
}
\value{
data.frame with columns \code{t}, \code{day}, and \code{dayMon} to \code{daySun}
}
\description{
Create data frame with day of week indicators
}
\details{
Takes the output of \code{points_to_grid()} and generates a data frame with indicator
variables for each day of the week for use in the \code{lgcp_fit()} function.
}
\examples{
b1 <- st_sf(st_sfc(st_polygon(list(cbind(c(0,3,3,0,0),c(0,0,3,3,0))))))
g1 <- create_grid(b1,0.5)
dp <- data.frame(y=runif(10,0,3),x=runif(10,0,3),date=paste0("2021-01-",11:20))
dp <- create_points(dp,pos_vars = c('y','x'),t_var='date')
g1 <- points_to_grid(g1, dp, laglength=5)
get_dow(g1)
}
