% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_Lu}
\alias{get_Lu}
\title{Generates the random effect terms from independent samples}
\usage{
get_Lu(L, u)
}
\arguments{
\item{L}{Lower triangular matrix representing the Cholesky decomposition of the spatial terms in one time period.}

\item{u}{A matrix of samples of u}
}
\value{
A matrix Lu.
}
\description{
For the MCMCML algorithms the spatial innovation terms are v. The Cholesky decomposition of the covariance matrix
of v is L. Random effects are sampled as standard normals u~N(0,I). This function generates v from samples u.
}
\details{
The LGCP model specified in this package uses an auto-regressive specification for the random effect
terms. If S(t) are the random effects for time period t, then these are specified as S(t) = rho*S(t-1) + v(t)
where v(t) is a spatial innovation term specified as a spatial Gaussian process. The Cholesky decomposition
of the covariance of v(t) is L and is the same in each time period, such that Lu = v where u~N(0,I). This function
produces v from L and u. The MCMCML algorithms also generate multiple samples of u so multiple samples of v are returned.
}
