% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_ZLu}
\alias{get_ZLu}
\title{Generates samples of the latent field from samples of u}
\usage{
get_ZLu(L, u, nT, rho)
}
\arguments{
\item{L}{Lower triangular matrix representing the Cholesky decomposition of the spatial terms in one time period.}

\item{u}{A matrix of samples of u}

\item{nT}{Integer specifying the number of time periods}

\item{rho}{Scalar, the auto-regressive parameter value.}
}
\value{
A matrix ZL
}
\description{
For the MCMCML algorithms the matrice Z and L multiply the random effects u~N(0,I) to produce
the latent field.
}
\details{
The LGCP model specified in this package uses an auto-regressive specification for the random effect
terms. If S(t) are the random effects for time period t, then these are specified as S(t) = rho*S(t-1) + v(t)
where v(t) is a spatial innovation term specified as a spatial Gaussian process. The Cholesky decomposition
of the covariance of v(t) is L and is the same in each time period, such that Lu = v where u~N(0,I). The matrix
Z generates the weighted sums of v(t) such that S = ZLu. This function produces S given Z, L, and u.
}
