% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{inv_ldlt}
\alias{inv_ldlt}
\title{Generates an approximate Cholesky decomposition of the covariance matrix}
\usage{
inv_ldlt(A, D, NN)
}
\arguments{
\item{A}{A "collapsed" matrix A output from \link[rts2]{get_AD}}

\item{D}{A diagonal matrix for the approximate LDL decomposition from \link[rts2]{get_AD}.}

\item{NN}{A matrix of integers with rows equal to the number of nearest neighbours and columns equal to the number of observations.
Each element is the set of nearest neighbours. See \link[rts2]{genNN}.}
}
\value{
A lower triangular matrix
}
\description{
Takes the output from the function \link[rts2]{get_AD} and generates the approximate
Cholesky decomposition of the covariance matrix of the random effects.
}
\details{
For the nearest neighbour Gaussian process approximation, the covariance matrix of the random effects/
latent fieldis approximated by \eqn{\Sigma = (I-A)^{-T}D(I-A)^{-1}} where matrices
A and D are determined by the set of nearest neighbours in \code{NN} (see \link[rts2]{genNN}).
The matrices A and D are returned by the function \link[rts2]{get_AD}.
Many of the calculations for the maximum likelihood model fitting methods
require the Cholesky decomposition of the covariance matrix \eqn{\Sigma = LL^T}, which is approximated here
by \eqn{(I-A)^{-T}D^{1/2}}. This function generates this lower triangular matrix
using a forward substitution algorithm as A is strictly lower triangular and D is diagonal.
}
