% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcml_functions.R
\name{lgcp_mcmcml}
\alias{lgcp_mcmcml}
\title{MCMCML Algorithm for the LGCP}
\usage{
lgcp_mcmcml(
  y,
  X,
  coords,
  popdens,
  nT,
  region_data,
  start,
  mod = "exp",
  use_cmdstanr = TRUE,
  mcml_options = list(useNN = FALSE, mcnr = FALSE, known_theta = FALSE, trace = 1, nNN =
    10, tol = 0.01, maxiter = 10),
  mcmc_options = list(warmup = 100, sampling = 100),
  verbose = TRUE
)
}
\arguments{
\item{y}{Vector of outcome counts}

\item{X}{Matrix of covariates for the linear predictor}

\item{coords}{A data frame providing the x and y coordinates of the computational grid}

\item{popdens}{Vector with the log population density}

\item{nT}{Number of time periods}

\item{region_data}{Optional. A list with the named elements \code{ncell}, \code{cell_id}, and \code{q_weights} giving, respectively
the number of cells overlapping each region, the indexes of the cells overlapping each region in order, and the weights
as the proportion of the area of the cell in each overlapping cell in order.}

\item{start}{Starting values of the model parameteters in the order c(beta, theta, rho). Rho is only for models with nT>1}

\item{mod}{Either "exp" for exponential or "sqexp" for squared exponential covariance function}

\item{use_cmdstanr}{Logical indicating whether to use cmdstanr}

\item{mcml_options}{List of options for the algorithm. See details.}

\item{mcmc_options}{List of options for the MCMC sampling. See details.}

\item{verbose}{Logical indicating whether to provide detailed feedback.}
}
\value{
A named list with the estimated parameters, number of iterations, whether the algorithm converged, and
the random effect samples.
}
\description{
Processes data and runs Markov Chain Monte Carlo Maximum Likelihood algorithms to fit
Log Gaussian Cox Process (LGCP) model. Used internally. Runs NNGP and full LGCP and both
standard and "region" models.
}
\details{
The argument \code{mcml_options} is a named list with the options of whether to use NNGP (\code{useNN}), whether to use
Newton-Raphson (\code{mcnr}), whether to treat the covariance parameters as known (\code{known_theta}), whether to provide
more detailed output (\code{trace}), the number of nearest neighbours if using NNGP (\code{nNN}), the tolerance for when to
terminate the algorithm (\code{tol}), and the maximum number of algorithm iterations (\code{maxiter}). The argument
\code{mcmc_options} is a named list with the number of warmup and sampling iterations for the MCMC sampler.
}
