% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lgcp_optim}
\alias{lgcp_optim}
\title{Maximum likelihood optimisation step for MCMCML algorithm for full LGCP model}
\usage{
lgcp_optim(
  cov,
  data,
  X,
  y,
  u,
  nT,
  start,
  offset,
  trace,
  mcnr = TRUE,
  known_cov = FALSE
)
}
\arguments{
\item{cov}{An integer matrix with columns of block identifier, dimension of block, function definition, number of variables
in the argument to the funciton, and index of the parameters, respectively. Rows are specific functions of each block.}

\item{data}{Vector of data. Created by flattening the matrices in column-major order of the data used in each block.}

\item{X}{Matrix of covariates for the linear predictor}

\item{y}{Vector of outcome counts for the computational grid cells}

\item{u}{Matrix of samples of the random effects terms with rows equal to the number of grid cells times by number of time
periods, and columns equal to the number of samples.}

\item{nT}{Integer specifying the number of time periods}

\item{start}{Vector of starting values for the model parameters in order beta, theta (covariance parameters), rho}

\item{offset}{Vector of offset values}

\item{trace}{Either 0, 1, or 2 indicating the level of detail to print to the console}

\item{mcnr}{Logical indicating whether to use newton-raphson (\code{TRUE}) or expectation maximisation (\code{FALSE}) for the estimation
of the parameters of the linear predictor}

\item{known_cov}{Logical indicating whether to treat the covariance parameters as known.}
}
\value{
A named list with beta, theta, and rho parameters
}
\description{
Maximum likelihood optimisation step for linear predictor parameters and covariance
parameters for MCMCML algorithm. The model is a point data full LGCP. This function is
used internally, use \code{lgcp_fit_ml()} function in the \link[rts2]{grid} class.
}
