% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{nngp_optim_la}
\alias{nngp_optim_la}
\title{Maximum likelihood model fitting using Laplace approximation to likelihood and nearest neighbour Gaussian process approximation}
\usage{
nngp_optim_la(
  cov,
  data,
  X,
  y,
  NN,
  start,
  offset,
  nT,
  known_cov = FALSE,
  usehess = FALSE,
  tol = 0.001,
  verbose = TRUE,
  trace = 0L,
  maxiter = 10L
)
}
\arguments{
\item{cov}{An integer matrix with columns of block identifier, dimension of block, function definition, number of variables
in the argument to the funciton, and index of the parameters, respectively. Rows are specific functions of each block.}

\item{data}{Vector of data. Created by flattening the matrices in column-major order of the data used in each block.}

\item{X}{Matrix of covariates for the linear predictor}

\item{y}{Vector of outcome counts for the computational grid cells}

\item{NN}{A matrix of integers with rows equal to the number of nearest neighbours and columns equal to the number of observations.
Each element is the set of nearest neighbours. See \link[rts2]{genNN}.}

\item{start}{Vector of starting values for the model parameters in order beta, theta (covariance parameters), rho}

\item{offset}{Vector of offset values}

\item{nT}{Integer specifying the number of time periods}

\item{known_cov}{Logical indicating whether to treat the covariance parameters as known.}

\item{usehess}{Logical indicating whether to use the Hessian matrix to estimate standard errors}

\item{tol}{Maximum difference between parameter values between iterations at which the algorithm is considered to have converged}

\item{verbose}{Logical indicating whether to provide output to the console}

\item{trace}{Either 0, 1, or 2 indicating the level of detail to print to the console}

\item{maxiter}{Integer. Maximum number of iterations of the algorithm.}
}
\value{
A named list with beta, theta, and rho parameters with standard errors and random effect estimates
}
\description{
Model fitting of the full Log Gaussian Cox Process model using a Laplace approximation to the likelihood.
This function is called internally, see the \code{fit_lgcp_la} in the \link[rts2]{grid} class.
}
