% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_mcmlrts.R
\name{predict.rtsFit}
\alias{predict.rtsFit}
\title{Predict from a \code{rtsFit} object}
\usage{
\method{predict}{rtsFit}(object, ...)
}
\arguments{
\item{object}{A \code{rtsFit} object.}

\item{...}{Further arguments passed from other methods}
}
\value{
Nothing. Called for effects.
}
\description{
Predictions cannot be generated directly from an \code{rtsFit} object, rather new predictions should be
generated using the original \code{grid} object. A message is printed to the user.
}
