% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_mcmlrts.R
\name{print.mcmlrts}
\alias{print.mcmlrts}
\title{Prints an mcmlrts fit output}
\usage{
\method{print}{mcmlrts}(x, ...)
}
\arguments{
\item{x}{an object of class "\code{mcmlrts}" as a result of a call to fit_ml}

\item{...}{Further arguments passed from other methods}
}
\value{
No return value, called for side effects.
}
\description{
Print method for class "\code{mcmlrts}"
}
\details{
\code{print.mcmlrts} tries to replicate the output of other regression functions, such
as \code{lm} and \code{lmer} reporting parameters, standard errors, and z- and p- statistics.
The z- and p- statistics should be interpreted cautiously however, as generalised
linear mixed models can suffer from severe small sample biases where the effective
sample size relates more to the higher levels of clustering than individual observations.
}
