% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_mcmlrts.R
\name{print.rtsFit}
\alias{print.rtsFit}
\title{Prints an rtsFit fit output}
\usage{
\method{print}{rtsFit}(x, ...)
}
\arguments{
\item{x}{an object of class "\code{rtsFit}"}

\item{...}{Further arguments passed from other methods}
}
\value{
No return value, called for side effects.
}
\description{
Print method for class "\code{rtsFit}"
}
\details{
\code{print.rtsFit} tries to replicate the output of other regression functions, such
as \code{lm} and \code{lmer} reporting parameters, standard errors, and z- and p- statistics for maximum
likelihood esitmates, or posterior means, standard deviations and credible intervals for
Bayesian models.
}
