% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_mcmlrts.R
\name{print.rtsFitSummary}
\alias{print.rtsFitSummary}
\title{Prints an rtsFitSummary fit output}
\usage{
\method{print}{rtsFitSummary}(x, ...)
}
\arguments{
\item{x}{an object of class "\code{rtsFitSummary}"}

\item{...}{Further arguments passed from other methods}
}
\value{
No return value, called for side effects.
}
\description{
Print method for class "\code{rtsFitSummary}"
}
\details{
\code{print.rtsFitSummary} prints the summary of an rtsFit, see \link[rts2]{summary.rtsFit}
}
