% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_mcmlrts.R
\name{random.effects}
\alias{random.effects}
\title{Extracts the random effect estimates}
\usage{
random.effects(object)
}
\arguments{
\item{object}{An \code{mcml} model fit.}
}
\value{
A matrix of dimension (number of fixed effects ) x (number of MCMC samples). For Laplace approximation, the number of "samples" equals one.
}
\description{
Extracts the random effect estimates or samples from an \verb{rtsFit object} returned from call of \code{lgcp_ml()} or \code{lgcp_bayes()} in the \link[rts2]{grid} class.
}
