% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{region_intensity}
\alias{region_intensity}
\title{Generates predictions of the intensity of each region for the region model}
\usage{
region_intensity(xb, y, zu, offset, nT, n_cell, cell_id, q_weights)
}
\arguments{
\item{xb}{A vector of values of the linear predictor X*beta}

\item{y}{Vector of outcome counts}

\item{zu}{Matrix with samples of the random field, see \link[rts2]{get_ZLu}.}

\item{offset}{Vector of offset values}

\item{nT}{Integer. Number of time periods.}

\item{n_cell}{Vector of integers specifying the number of grid cells intersecting each region,
see the \code{region_data()} function in the \link[rts2]{grid} class.}

\item{cell_id}{Vector of indices of grid cells overlapping each region, in order,
see the \code{region_data()} function in the \link[rts2]{grid} class.}

\item{q_weights}{Vector specifying the proportion of the area of each region covered by overlapping
grid cells, see the \code{region_data()} function in the \link[rts2]{grid} class.}
}
\value{
A matrix with samples of the intensity in each region at each time point.
}
\description{
The "region model" specifies the intensity for each region as a combination of the
offset, linear predictor, and a weighted combination of the intensity of overlapping
computational grid cells. This function returns the region intensities given above
components.
}
