% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_mcmlrts.R
\name{residuals.grid}
\alias{residuals.grid}
\title{Residuals method for a \code{grid} object}
\usage{
\method{residuals}{grid}(object, type, ...)
}
\arguments{
\item{object}{A \code{grid} object.}

\item{type}{Either "standardized" or "raw"}

\item{...}{Further arguments passed from other methods}
}
\value{
A matrix with number of columns corresponding to the number of MCMC samples.
}
\description{
Conditional raw or standardised residuals are returned for a stored \code{rtsFit} objects. If no prior model fit is stored,
then an error is returned.
}
