% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_mcmlrts.R
\name{residuals.rtsFit}
\alias{residuals.rtsFit}
\title{Residuals method for a \code{rtsFit} object}
\usage{
\method{residuals}{rtsFit}(object, type, ...)
}
\arguments{
\item{object}{A \code{rtsFit} object.}

\item{type}{Either "standardized" or "raw"}

\item{...}{Further arguments passed from other methods}
}
\value{
A matrix with number of columns corresponding to the number of MCMC samples.
}
\description{
Conditional raw or standardised residuals for \code{rstFit} objects. The residuals are limited to conditional raw or standardised
residuals currently to avoid copying the often large amount of model data stored in the associated grid object.
}
