% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcml_functions.R
\name{sample_u}
\alias{sample_u}
\title{Generates samples of the random effects}
\usage{
sample_u(
  y,
  xb,
  coords,
  nT,
  region_data,
  start,
  mod = "exp",
  use_cmdstanr = TRUE,
  verbose = TRUE,
  mcmc_options = list(warmup = 100, sampling = 100)
)
}
\arguments{
\item{y}{Vector of outcome counts}

\item{xb}{Vector of values of the linear predictor including any offsets}

\item{coords}{A data frame providing the x and y coordinates of the computational grid}

\item{nT}{Number of time periods}

\item{region_data}{Optional. A list with the named elements \code{ncell}, \code{cell_id}, and \code{q_weights} giving, respectively
the number of cells overlapping each region, the indexes of the cells overlapping each region in order, and the weights
as the proportion of the area of the cell in each overlapping cell in order.}

\item{start}{Values of the covariance parameters in the order c(theta, rho). Rho is only for models with nT>1}

\item{mod}{Either "exp" for exponential or "sqexp" for squared exponential covariance function}

\item{use_cmdstanr}{Logical indicating whether to use cmdstanr}

\item{mcmc_options}{List of options for the MCMC sampling, specifically the number of warmup and sampling iterations.}
}
\value{
A matrix of samples
}
\description{
Using known model parameters generates samples of the random effects from either the standard or regional model
}
