% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{scale_conversion_factor}
\alias{scale_conversion_factor}
\title{Returns scale conversion factor}
\usage{
scale_conversion_factor(grid_data)
}
\arguments{
\item{grid_data}{sf object. See \link[rts2]{create_grid}}
}
\value{
numeric
}
\description{
Coordinates are scaled to \verb{[-1,1]} for \link[rts2]{lgcp_fit}. This function
returns the scaling factor for this conversion.
}
\examples{
b1 = sf::st_sf(sf::st_sfc(sf::st_polygon(list(cbind(c(0,3,3,0,0),c(0,0,3,3,0))))))
g1 <- create_grid(b1,0.5)
scale_conversion_factor(g1)
}
