% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_mcmlrts.R
\name{summary.grid}
\alias{summary.grid}
\title{Summarizes a \code{grid} object}
\usage{
\method{summary}{grid}(object, ...)
}
\arguments{
\item{object}{A \code{grid} object.}

\item{...}{Further arguments passed from other methods}
}
\value{
Nothing. Called for effects.
}
\description{
Summarizes \code{grid} object.
}
