% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_mcmlrts.R
\name{summary.mcmlrts}
\alias{summary.mcmlrts}
\title{Summarises an mcmlrts fit output}
\usage{
\method{summary}{mcmlrts}(object, ...)
}
\arguments{
\item{object}{an object of class "\code{mcmlrts}" as a result of a call to fit_ml}

\item{...}{Further arguments passed from other methods}
}
\value{
A list with random effect names and a data frame with random effect mean and credible intervals
}
\description{
Summary method for class "\code{mcmlrts}"
}
\details{
\code{print.mcmlrts} tries to replicate the output of other regression functions, such
as \code{lm} and \code{lmer} reporting parameters, standard errors, and z- and p- statistics.
The z- and p- statistics should be interpreted cautiously however, as generalised
linear miobjected models can suffer from severe small sample biases where the effective
sample size relates more to the higher levels of clustering than individual observations.
}
