% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_mcmlrts.R
\name{summary.rtsFit}
\alias{summary.rtsFit}
\title{Summary method for class "rtsFit"}
\usage{
\method{summary}{rtsFit}(object, ...)
}
\arguments{
\item{object}{an object of class "\code{rtsFit}" as a result of a call to \code{lgcp_ml()} or \code{lgcp_bayes()}}

\item{...}{Further arguments passed from other methods}
}
\value{
An rtsFitSummary object
}
\description{
Summary method for class "\code{rtsFit}"
}
\details{
The summary methods aims to replicate the output of other regression model fitting functions and reports
central point estimates, relevant test statistics, and uncertainty intervals. In addition, the returned
summary object will also include time period specific relative risk and incidence predictions.
}
