% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_mcmlrts.R
\name{vcov.grid}
\alias{vcov.grid}
\title{Calculate Variance-Covariance matrix for a maximum likelihood object stored in \code{grid}}
\usage{
\method{vcov}{grid}(object, ...)
}
\arguments{
\item{object}{A \code{grid} object.}

\item{...}{Further arguments passed from other methods}
}
\value{
A variance-covariance matrix.
}
\description{
Returns the variance-covariance matrix for a LGCP object fit using maximum likelihood. If no relevant
model is stored then the function returns an error
}
