% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_mcmlrts.R
\name{vcov.rtsFit}
\alias{vcov.rtsFit}
\title{Extract the Variance-Covariance matrix for a \code{rtsFit} object}
\usage{
\method{vcov}{rtsFit}(object, ...)
}
\arguments{
\item{object}{A \code{rtsFit} object.}

\item{...}{Further arguments passed from other methods}
}
\value{
A variance-covariance matrix.
}
\description{
Returns the calculated variance-covariance matrix for a \code{rtsFit} object that was fit using maximum likelihood methods.
Bayesian models will return an error.
}
