% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_mcmlrts.R
\name{coef.rtsFit}
\alias{coef.rtsFit}
\title{Extracts coefficients from a mcml object}
\usage{
\method{coef}{rtsFit}(object, ...)
}
\arguments{
\item{object}{An \code{mcml} model fit.}

\item{...}{Further arguments passed from other methods}
}
\value{
A data frame summarising the parameters including the random effects.
}
\description{
Extracts the coefficients from an \code{mcml} object returned from a call of \code{MCML} or \code{LA} in the \link[glmmrBase]{Model} class.
}
