% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_mcmlrts.R
\name{logLik.rtsFit}
\alias{logLik.rtsFit}
\title{Extracts the log-likelihood from an rtsFit object}
\usage{
\method{logLik}{rtsFit}(object, ...)
}
\arguments{
\item{object}{An \code{rtsFit} model fit.}

\item{...}{Further arguments passed from other methods}
}
\value{
An object of class \code{logLik} for maximum likelihood model fits, otherwise it returns an error.
}
\description{
Extracts the final log-likelihood value from an rtsFit object. Only returns a value for maximum likelihood
model fits, otherwise it produces an error.
}
