% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rts2-package.R
\docType{package}
\name{rts2-package}
\alias{rts2-package}
\alias{rts2}
\title{The 'rts2' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (2020). RStan: the R interface to Stan. R package version 2.21.2. https://mc-stan.org
}
